% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sluPlaces.R
\docType{data}
\name{sluPlaces}
\alias{sluPlaces}
\title{Campus Locations at Saint Louis University}
\format{A data frame with 106 rows and 7 variables:
\describe{
  \item{id}{record id number}
  \item{name}{location name}
  \item{lng}{longitude (decimal degrees)}
  \item{lat}{latitude (decimal degrees)}
  }}
\usage{
data(sluPlaces)
}
\description{
A simple data set containing a few locations on SLU's North Campus.
}
\examples{
str(sluPlaces)
head(sluPlaces)

}
\keyword{datasets}
