% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_as_sf.R
\name{stl_as_sf}
\alias{stl_as_sf}
\title{Load tibble as sf}
\usage{
stl_as_sf(table)
}
\arguments{
\item{table}{Unquoted name of one of the accepted tables}
}
\value{
A simple features object with the specified tibble joined to the appropriate geometry.
}
\description{
For specified tibbles, this function will load them as simple features with appropriate
    geometric data.
}
\details{
This function currently works with the tibbles that contain census tract level data:
    stl_tbl_asthma, stl_tbl_income, stl_tbl_insurance, stl_tbl_lead, and stl_tbl_smoking
}
