% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_sf_boundary.R
\docType{data}
\name{stl_sf_boundary}
\alias{stl_sf_boundary}
\title{City of St. Louis Boundary}
\format{A simple feature with 1 rows and 5 variables:
\describe{
  \item{GEOID}{county FIPS code}
  \item{NAME}{polygon name}
  \item{LAND_SQMI}{land area of city, square miles}
  \item{WATER_SQMI}{water area of city, square miles}
  \item{geometry}{simple features geometry}
  }}
\usage{
data(stl_sf_boundary)
}
\description{
A simple features data set containing a single polygon with the extent of the City of St. Louis's boundary.
}
\examples{
class(stl_sf_boundary)
str(stl_sf_boundary)
head(stl_sf_boundary)

}
\keyword{datasets}
