% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_sf_historic.R
\docType{data}
\name{stl_sf_historic}
\alias{stl_sf_historic}
\title{Historic Districts in the City of St. Louis}
\format{A simple feature with 102 rows and 6 variables:
\describe{
  \item{NUMBER}{District number}
  \item{NAME}{District name}
  \item{LAYER}{District type}
  \item{COUNT}{count}
  \item{DATE}{District creation date}
  \item{geometry}{simple features geometry}
  }}
\usage{
data(stl_sf_historic)
}
\description{
A simple features data set containing polygons for the major bodies of water in the city.
}
\examples{
class(stl_sf_historic)
str(stl_sf_historic)
head(stl_sf_historic)

}
\keyword{datasets}
