% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_sf_hydro.R
\docType{data}
\name{stl_sf_hydro}
\alias{stl_sf_hydro}
\title{Major Bodies of Water in the City of St. Louis}
\format{A simple feature with 2 rows and 4 variables:
\describe{
  \item{HYDROID}{Census identification number}
  \item{NAME}{name of body of water}
  \item{SQMETER}{area covered by water, square meters}
  \item{geometry}{simple features geometry}
  }}
\usage{
data(stl_sf_hydro)
}
\description{
A simple features data set containing polygons for the major bodies of water in the city.
}
\examples{
class(stl_sf_hydro)
str(stl_sf_hydro)
head(stl_sf_hydro)

}
\keyword{datasets}
