% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_tbl_sluPlaces.R
\docType{data}
\name{stl_tbl_sluPlaces}
\alias{stl_tbl_sluPlaces}
\title{Campus Locations at Saint Louis University}
\format{A tibble with 6 rows and 4 variables:
\describe{
  \item{id}{record id number}
  \item{name}{location name}
  \item{lng}{longitude (decimal degrees)}
  \item{lat}{latitude (decimal degrees)}
  }}
\usage{
data(stl_tbl_sluPlaces)
}
\description{
A simple data set containing a few locations on SLU's North Campus.
}
\examples{
str(stl_tbl_sluPlaces)
head(stl_tbl_sluPlaces)

}
\keyword{datasets}
