% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl_tbl_water.R
\docType{data}
\name{stl_tbl_water}
\alias{stl_tbl_water}
\title{Clean Water Act Listed Rivers and Streams}
\format{A tibble with 179 rows and 6 variables:
\describe{
  \item{YR}{year listed}
  \item{WBID}{water body id number}
  \item{WATER_BODY}{water body name}
  \item{POLLUTANT}{pollutant type}
  \item{SOURCE_}{pollutant source}
  \item{PERM_ID}{permanent id number}
  }}
\usage{
data(stl_tbl_water)
}
\description{
A data set containing rivers and streams listed under the clean water act in St. Louis.
}
\examples{
str(stl_tbl_water)
head(stl_tbl_water)
summary(stl_tbl_water$pctUnins)

}
\keyword{datasets}
