% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlBoundary.R
\docType{data}
\name{stlBoundary}
\alias{stlBoundary}
\title{City of St. Louis Boundary}
\format{A simple feature with 1 rows and 6 variables:
\describe{
  \item{LAYER}{polygon description}
  \item{AREA_SQMI}{square mileage of city}
  \item{Name}{polygon name}
  \item{Shape_Leng}{length of polygon}
  \item{Shape_Area}{area of polygon}
  \item{geometry}{simple features geometry}
  }}
\usage{
data(stlBoundary)
}
\description{
A simple features data set containing a single polygon with the extent of the City of St. Louis's boundary.
}
\examples{
class(stlBoundary)
str(stlBoundary)
head(stlBoundary)

}
\keyword{datasets}
