% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stldata-package.r
\docType{package}
\name{stldata}
\alias{stldata}
\alias{stldata-package}
\title{stldata: City of St. Louis Data}
\description{
This package contains various datasets representing the City of St. Louis.
These datasets are primary designed for teaching statistics, data science, and
spatial data analysis using R.
}
