% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlHydro.R
\docType{data}
\name{stlHydro}
\alias{stlHydro}
\title{Major Bodies of Water in the City of St. Louis}
\format{A simple feature with 2 rows and 11 variables:
\describe{
  \item{ANSICODE}{ANSI identification number}
  \item{HYDROID}{Census identification number}
  \item{FULLNAME}{name of body of water}
  \item{ALAND}{area covered by land}
  \item{INTPTLAT}{interpolated latitude of polygon's centroid}
  \item{INTPTLON}{interpolated longitude of polygon's centroid}
  \item{Shape_Leng}{length of polygon}
  \item{Shape_Area}{area of polygon}
  \item{geometry}{simple features geometry}
  }}
\usage{
data(stlHydro)
}
\description{
A simple features data set containing polygons for the major bodies of water in the city.
}
\examples{
class(stlHydro)
str(stlHydro)
head(stlHydro)

}
\keyword{datasets}
