% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlInsurance.R
\docType{data}
\name{stlInsurance}
\alias{stlInsurance}
\title{Crude Prevalance of No Health Insurance, 2014}
\format{A data frame with 106 rows and 7 variables:
\describe{
  \item{geoID}{full census tract id number}
  \item{tractCE}{census tract id number}
  \item{nameLSAD}{census tract string}
  \item{pctUnins}{crude prevalance estimate of current no health insurance, percent}
  \item{pctUnins_Low}{lower bound of crude prevalance estimate}
  \item{pctUnins_High}{upper bound of crude prevalance estimate}
  }}
\source{
\href{https://www.cdc.gov/500cities/}{Centers for Disease Control 500 Cities Project}

#' @examples
str(stlInsurance)
head(stlInsurance)
summary(stlInsurance$pctUnins)
}
\usage{
data(stlInsurance)
}
\description{
A data set containing the estimated prevalance of no health insurance among City of St. Louis residents by Census Tract.
}
\keyword{datasets}
