% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlSmoking.R
\docType{data}
\name{stlSmoking}
\alias{stlSmoking}
\title{Crude Prevalance of Current Smoking, 2014}
\format{A data frame with 106 rows and 7 variables:
\describe{
  \item{geoID}{full census tract id number}
  \item{tractCE}{census tract id number}
  \item{nameLSAD}{census tract string}
  \item{pctSmoke}{crude prevalance estimate of current no health insurance, percent}
  \item{pctSmoke_Low}{lower bound of crude prevalance estimate}
  \item{pctSmoke_High}{upper bound of crude prevalance estimate}
  }}
\source{
\href{https://www.cdc.gov/500cities/}{Centers for Disease Control 500 Cities Project}

#' @examples
str(stlSmoking)
head(stlSmoking)
summary(stlSmoking$pctSmoke)
}
\usage{
data(stlSmoking)
}
\description{
A data set containing the estimated prevalance of current smoking among City of St. Louis residents by Census Tract.
}
\keyword{datasets}
