% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stlData-package.R
\docType{package}
\name{stldata}
\alias{stldata}
\alias{stldata-package}
\title{City of St. Louis Data}
\description{
This package contains various datasets representing various phenomena in the City of St. Louis.
These datasets are primary designed for teaching statistics, data science, and
spatial data analysis using R. It currently contains five datasets describing murders
in the City, blood lead level test results among children, and a number of data sets from
the Centers for Disease Control's 500 Cities Project. If you have additional
data you would like to donate to the package or variables you would like to see added to
a table, please contact Chris by opening an issue on the package's
\href{https://github.com/slu-openGIS/stldata/issues}{Github page}.
}
